Unicode support!  Well, after a fashion:
Currently python_reader makes no attempt to work out the correct
encoding and UnixConsole defaults to 'utf-8'.  This is wrong, but it works
for me.

Reader.get_buffer returns a unicode string.  Ilpo & I decided a while back
that this is the wrong thing to do.

Fixes for both should follow shortly.


git-svn-id: http://codespeak.net/svn/pyrepl/trunk/pyrepl@4624 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada